一覧画面から詳細ページに画面遷移するように調整
