+++ app/javascript/components/Header.vue
@@ -7,6 +7,10 @@
     <v-toolbar-title> {{ title }} </v-toolbar-title>
 
     <v-spacer></v-spacer>
+    <v-btn>
+      新規登録
+    </v-btn>
+    <v-spacer></v-spacer>
 
     <v-btn icon>
       <v-icon>fas fa-search</v-icon>
