+++ app/javascript/components/PostShow.vue
@@ -1,0 +1,5 @@
+<template>
+  <div>
+    <p>post_id : {{ $route.params.id }} の画面です</p>
+  </div>
+</template>
