+++ app/javascript/router/router.js
@@ -1,7 +1,8 @@
 import Vue from "vue";
 import Router from "vue-router";
 
-import ArticleList from "../components/ArticleList";
+import PostList from "../components/PostList";
+import PostShow from "../components/PostShow";
 
 Vue.use(Router);
 
@@ -10,7 +11,12 @@ const router = new Router({
   routes: [
     {
       path: "/",
-      component: ArticleList,
+      component: PostList,
+    },
+    {
+      path: "/posts/:id",
+      name: "PostShow",
+      component: PostShow,
     },
   ],
 });
